requirements.txt
selenium
webdriver_manager
openpyxl
pandas
python-dotenv